- Remove event handlers from IScopeManager interface and ScopeManagerBase. Also remove SpanEventArgs - Remove ScopeManagerBase and move all logic to AsyncLocalScopeManager
