+++ tracer/src/Datadog.Trace/IScopeManager.cs
@@ -12,18 +12,6 @@ namespace Datadog.Trace
     /// </summary>
     internal interface IScopeManager
     {
-        event EventHandler<SpanEventArgs> TraceStarted;
-
-        event EventHandler<SpanEventArgs> SpanOpened;
-
-        event EventHandler<SpanEventArgs> SpanActivated;
-
-        event EventHandler<SpanEventArgs> SpanDeactivated;
-
-        event EventHandler<SpanEventArgs> SpanClosed;
-
-        event EventHandler<SpanEventArgs> TraceEnded;
-
         Scope Active { get; }
 
         Scope Activate(Span span, bool finishOnClose);
